api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
pydantic==1.10.13
jinja2==3.1.2
websockets==11.0.3
python-dotenv==1.0.0